refactor(governance): move invite lifecycle routes
